Game sudoku cell field, showing cell number or notes

+++ examples/minesweeper/app/scenes/scene_game/game.go
@@ -15,7 +15,7 @@ type Game struct {
 
 func newGame(r, c, m int) *Game {
 	g := &Game{}
-	g.layout = eui.NewGridLayoutRightDown(r, c)
+	g.layout = eui.NewGridLayoutRightDown(float64(r), float64(c))
 	g.layout.SetCellMargin(1)
 	g.field = game.NewMinedField(r, c, m)
 	g.field.State.Attach(g)
